# openapi.yaml
openapi: 3.0.3
info:
  title: XBP Monitoring API
  version: "1.0.0"
  description: |
    XBP Monitoring is a synthetic monitoring service that continuously probes HTTP endpoints and executes multi-step stories.
    
    ## Features
    - **Probes**: Single HTTP endpoint checks with configurable expectations
    - **Stories**: Multi-step workflows that can pass data between steps
    - **Results**: Historical data of all probe and story executions
    - **Metrics**: Prometheus-compatible metrics endpoint for observability
    
    ## Authentication
    This API does not require authentication for most endpoints. The reload endpoint requires a token header (`x-xbp-reload-token`) matching the `XBP_RELOAD_TOKEN` environment variable.
    
    ## Rate Limiting
    No rate limiting is currently enforced. Trigger endpoints execute probes/stories synchronously.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://{host}:{port}
    description: Application server
    variables:
      host:
        default: api
      port:
        default: "3000"
  - url: http://{host}:{port}
    description: Prometheus metrics server (when enabled)
    variables:
      host:
        default: metrics
      port:
        default: "9464"
security: []
...
paths:
...
  /-/reload:
    post:
      tags:
        - Admin
      summary: Reload configuration and restart monitors
      description: |
        Reloads config (from `XBP_REMOTE_CONFIG_URL` if set, otherwise from the configured `--file`) and restarts all scheduled monitors.
        
        Requires header `x-xbp-reload-token` matching the `XBP_RELOAD_TOKEN` environment variable.
      operationId: reloadDaemon
      parameters:
        - name: x-xbp-reload-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reload succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReloadResponse"
        "403":
          description: Forbidden - missing or invalid reload token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Reload failed or token not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      description: Standard error response format returned for 4XX and 5XX status codes
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message describing what went wrong
          example: "Probe 'nonexistent-probe' not found"
...
    ReloadResponse:
      type: object
      required:
        - reloaded
        - probes
        - stories
      properties:
        reloaded:
          type: boolean
        probes:
          type: array
          items:
            type: string
        stories:
          type: array
          items:
            type: string